Fixed fix and fix and now ready
