+++ spec/spec_helper.rb
@@ -4,6 +4,11 @@ SimpleCov.start('rails')
 
 require 'capybara/rspec'
 
+# in spec/support/ and its subdirectories.
+Dir[Rails.root.join("spec/support/**/*.rb")].each { |f| require f }
+
+I18n.enforce_available_locales = false
+
 # This file was generated by the `rails generate rspec:install` command. Conventionally, all
 # specs live under a `spec` directory, which RSpec adds to the `$LOAD_PATH`.
 # The generated `.rspec` file contains `--require spec_helper` which will cause this
