+++ spec/views/exercises/index.html.erb_spec.rb
@@ -1,9 +1,9 @@
 require 'rails_helper'
 
+include RSpec::Mocks::ExampleMethods
+
 
 RSpec.describe "exercises/index", :type => :view do
-
-
 
   before(:each) do
 
@@ -16,22 +16,16 @@ RSpec.describe "exercises/index", :type => :view do
       )
     ])
 
-
    # allow(controller).to receive(:current_user).and_return(FactoryGirl.create(:user))
-
-
-
-
 
   end
 
 
-  #it "renders a list of exercises" do
+  it "renders a list of exercises again" do
     #controller.stub!(:current_user).and_return(User.create!(:name => "Koira"))
 
-
-   # render
-    #assert_select ".btn-exercise", :count => 2
-  #end
+    render
+    assert_select ".btn-exercise", :count => 2
+  end
 
 end
